+++ packages/react-composites/src/composites/CallComposite/components/buttons/People.tsx
@@ -2,22 +2,24 @@
 // Licensed under the MIT license.
 
 import React, { useMemo } from 'react';
-import { ControlBarButton, ControlBarButtonProps, ControlBarButtonStyles, useTheme } from '@internal/react-components';
-/* @conditional-compile-remove(rooms) */
-import { _usePermissions } from '@internal/react-components';
+import { ControlBarButton, ControlBarButtonStyles, useTheme } from '@internal/react-components';
 import { concatStyleSets } from '@fluentui/react';
 import { CallCompositeIcon } from '../../../common/icons';
 import { controlButtonBaseStyle } from '../../styles/Buttons.styles';
+import { CallControlOptions } from '../../types/CallControlOptions';
+import { useLocale } from '../../../localization/LocalizationProvider';
+import { isDisabled } from '../../utils';
 
 const icon = (): JSX.Element => <CallCompositeIcon iconName={'ControlButtonParticipants'} />;
 
 /**
  * @private
  */
-/** @beta */
-export const People = (props: ControlBarButtonProps): JSX.Element => {
-  const { strings, onRenderOnIcon, onRenderOffIcon, onClick } = props;
-
+export const People = (props: {
+  peopleButtonChecked?: boolean;
+  onPeopleButtonClicked?: () => void;
+  options?: CallControlOptions;
+}): JSX.Element => {
   const theme = useTheme();
   const styles: ControlBarButtonStyles = useMemo(
     () =>
@@ -27,22 +29,34 @@ export const People = (props: ControlBarButtonProps): JSX.Element => {
             background: theme.palette.neutralLight
           }
         },
-        props.styles ?? {},
         controlButtonBaseStyle
       ),
-    [props.styles, theme.palette.neutralLight]
+    [theme.palette.neutralLight]
+  );
+
+  const locale = useLocale();
+  // FIXME (?): Why is this using callWithChat strings?
+  const strings = useMemo(
+    () => ({
+      label: locale.strings.callWithChat.peopleButtonLabel,
+      tooltipOffContent: locale.strings.callWithChat.peopleButtonTooltipOpen,
+      tooltipOnContent: locale.strings.callWithChat.peopleButtonTooltipClose
+    }),
+    [locale]
   );
 
   return (
     <ControlBarButton
-      {...props}
       data-ui-id="call-composite-participants-button"
+      checked={props.peopleButtonChecked}
+      showLabel={props.options?.displayType !== 'compact'}
       strings={strings}
       labelKey={'peopleButtonLabelKey'}
-      onRenderOnIcon={onRenderOnIcon ?? icon}
-      onRenderOffIcon={onRenderOffIcon ?? icon}
-      onClick={onClick}
+      onRenderOnIcon={icon}
+      onRenderOffIcon={icon}
+      onClick={props.onPeopleButtonClicked}
       styles={styles}
+      disabled={isDisabled(props.options?.participantsButton)}
     />
   );
 };
